pyproject.toml
[project]
description = "Backend for Cabbo cab booking platform"
name = "cabbo"
requires-python = ">=3.10"
version = "0.1.0"

dependencies = [
  "alembic",
  "fastapi",
  "uvicorn",
  "python-dotenv",
  "sqlalchemy",
  "mysql-connector-python",
  "aiomysql",
  "email-validator",
  "twilio",
  "rich",
  "pydantic-settings",
  "sendgrid",
  "python-multipart",
  "mapbox",
  "Jinja2",
  "setuptools",
  "tzdata>=2025.2",
  "python-dateutil>=2.9.0.post0",
  "APScheduler>=3.10",
  "aiosmtplib>=5.0.0",
  "razorpay>=1.3.0",
]

[project.optional-dependencies]
dev = [
  "alembic>=1.18.0",
  # add other dev tools here, e.g. "pytest", "black", "mypy"
]

[tool.uv]
# uv-specific settings can be added here
